HangMan Game
This is my Rendition of the Famous Hang Man Game coded in Python! I employed some functions to facilitate the natural order of gameplay. I also added an Exception Handling Block just for the fun of it.
In this version of my game, you can choose the difficulty of the words you are given to try to solve!!!
All in-depth descriptions of the workings of the code can be seen in the comments I have labeled!

Have Fun playing!!!!!